requirements.txt
httpx
rich
pydantic
google-generativeai
python-dotenv
websockets
